rename model file to match defined class
